[#575] styling languages list
