+++ app/src/main/res/layout/language_list_item.xml
@@ -10,10 +10,11 @@
             android:id="@+id/language_checkbox"
             tools:text="English"
             android:clickable="false"
-            android:layout_marginLeft="16dp"
-            android:layout_marginStart="16dp"
+            android:layout_marginLeft="4dp"
+            android:layout_marginStart="4dp"
             android:paddingLeft="8dp"
             android:paddingStart="8dp"
+            android:layout_marginBottom="4dp"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"/>
 </LinearLayout>
