+++ app/src/main/res/layout/languages_list.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ListView xmlns:android="http://schemas.android.com/apk/res/android"
+          android:layout_width="match_parent"
+          android:layout_height="match_parent"
+          android:paddingTop="16dp"
+          android:paddingLeft="16dp"
+          android:paddingRight="16dp">
+</ListView>
